+++ docs/source/changelog.md
@@ -16,7 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.1.0] - 2024-01-XX
 
 ### Added
-- Initial release of spark-profiler
+- Initial release of pyspark-analyzer
 - Core `DataFrameProfiler` class for PySpark DataFrame analysis
 - Automatic computation of data type-specific statistics
 - Intelligent sampling for large datasets with `SamplingConfig`